+++ Plugin/Thunderstore/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 2.0.0
 - Stunnable
+- Can go outside (Config option)
+- Non-Deadly Interactions added
 
 ## 1.0.1
 - Patched for V50 release of Lethal Company.